+++ server/src/main/java/org/opensearch/telemetry/tracing/WrappedTracer.java
@@ -38,12 +38,17 @@ final class WrappedTracer implements Tracer {
     }
 
     @Override
-    public SpanScope startSpan(String spanName) {
+    public Span startSpan(SpanCreationContext context) {
+        return startSpan(context.getSpanName(), context.getAttributes());
+    }
+
+    @Override
+    public Span startSpan(String spanName) {
         return startSpan(spanName, Attributes.EMPTY);
     }
 
     @Override
-    public SpanScope startSpan(String spanName, Attributes attributes) {
+    public Span startSpan(String spanName, Attributes attributes) {
         return startSpan(spanName, (SpanContext) null, attributes);
     }
 
@@ -54,7 +59,22 @@ final class WrappedTracer implements Tracer {
     }
 
     @Override
-    public SpanScope startSpan(String spanName, SpanContext parentSpan, Attributes attributes) {
+    public ScopedSpan startScopedSpan(SpanCreationContext spanCreationContext) {
+        return startScopedSpan(spanCreationContext, null);
+    }
+
+    @Override
+    public ScopedSpan startScopedSpan(SpanCreationContext spanCreationContext, SpanContext parentSpan) {
+        return getDelegateTracer().startScopedSpan(spanCreationContext, parentSpan);
+    }
+
+    @Override
+    public SpanScope withSpanInScope(Span span) {
+        return getDelegateTracer().withSpanInScope(span);
+    }
+
+    @Override
+    public Span startSpan(String spanName, SpanContext parentSpan, Attributes attributes) {
         Tracer delegateTracer = getDelegateTracer();
         return delegateTracer.startSpan(spanName, parentSpan, attributes);
     }
@@ -70,7 +90,7 @@ final class WrappedTracer implements Tracer {
     }
 
     @Override
-    public SpanScope startSpan(String spanName, Map<String, List<String>> headers, Attributes attributes) {
+    public Span startSpan(String spanName, Map<String, List<String>> headers, Attributes attributes) {
         return defaultTracer.startSpan(spanName, headers, attributes);
     }
 }